// src/domain/enums/permission-action.ts
export enum PermissionAction {
  AGENTS_VIEW = "AGENTS_VIEW",
  AGENTS_WRITE = "AGENTS_WRITE",
  WABAS_VIEW = "WABAS_VIEW",
  WABAS_WRITE = "WABAS_WRITE",
  SERVICE_ISLANDS_VIEW = "SERVICE_ISLANDS_VIEW",
  SERVICE_ISLANDS_WRITE = "SERVICE_ISLANDS_WRITE",
  QUEUES_VIEW = "QUEUES_VIEW",
  QUEUES_WRITE = "QUEUES_WRITE",
  CONTACTS_VIEW = "CONTACTS_VIEW",
  CONTACTS_WRITE = "CONTACTS_WRITE",
  ACCESS_VIEW = "ACCESS_VIEW",
  ACCESS_WRITE = "ACCESS_WRITE",
  COMPANIES_MANAGE_OWN = "COMPANIES_MANAGE_OWN",
  COMPANIES_MANAGE_ALL = "COMPANIES_MANAGE_ALL",
}